LayoutTests/http/tests/media/video-cookie-httponly-expected.txt

EVENT(canplay)
Tests that the media player will send the relevant httponly cookies when requesting the media file.
